requirements.txt
gunicorn
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.2.0
playwright==1.51.0
propcache==0.3.0
pyee==12.1.1
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
